grpc_server/pyproject.toml
[project]
name = "graphiti-grpc-server"
version = "0.1.0"
description = "gRPC server for Graphiti - Temporally-aware Knowledge Graph for AI Agents"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
authors = [{ name = "Zep Software, Inc." }]
keywords = ["graphiti", "knowledge-graph", "grpc", "ai", "agents"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "grpcio-reflection>=1.60.0",
    "grpcio-health-checking>=1.60.0",
    "protobuf>=4.25.0",
    "graphiti-core>=0.23.0",
    "graphiti-server-common",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-instrumentation-grpc>=0.41b0",
    "falkordb>=1.4.0",
    "groq>=1.0.0",
    "voyageai>=0.3.7",
    "pyjwt>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "grpcio-testing>=1.60.0",
    "ruff>=0.1.0",
    "pyright>=1.1.350",
    "mypy-protobuf>=3.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"
exclude = ["src/generated"]

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.ruff.format]
quote-style = "single"

[tool.pyright]
pythonVersion = "3.10"
typeCheckingMode = "basic"
include = ["src"]
exclude = ["src/generated"]

[tool.uv.sources]
graphiti-core = { path = "../", editable = true }
graphiti-server-common = { path = "../server_common", editable = true }

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
filterwarnings = [
    "ignore:coroutine.*was never awaited:RuntimeWarning",
    "ignore::pydantic.warnings.PydanticDeprecatedSince20",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "src/generated/*",
    "*_pb2.py",
    "*_pb2_grpc.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
fail_under = 75
